Server/requirements.txt
flask
flask-cors
numpy
pillow
tensorflow==2.12.0
gunicorn
groq
python-dotenv